requirements.txt
absl-py==2.0.0
anyio==4.0.0
arandr==0.1.11
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.6.0
astroid==2.14.2
asttokens==2.2.1
async-lru==2.0.4
attrs==23.1.0
av==10.0.0
Babel==2.10.3
beautifulsoup4==4.11.2
bidict==0.22.1
bleach==6.1.0
blinker==1.5
certifi==2022.9.24
cffi==1.16.0
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
colorzero==2.0
comm==0.2.0
contourpy==1.2.0
cryptography==38.0.4
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.3.2
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distro==1.8.0
docutils==0.19
executing==2.0.1
fastjsonschema==2.19.0
filelock==3.13.1
Flask==2.2.2
Flask-SocketIO==5.3.6
flatbuffers==20181003210633
fonttools==4.44.3
fqdn==1.5.1
fsspec==2023.10.0
gpiozero==2.0
gTTS==2.4.0
gTTS-token==1.1.4
h11==0.14.0
html5lib==1.1
idna==3.3
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-metadata==4.12.0
imutils==0.5.4
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
isort==5.6.4
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.11.1
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.1
kiwisolver==1.4.5
lazy-object-proxy==1.9.0
lgpio==0.2.2.0
libevdev==0.5
logilab-common==1.9.8
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.8.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mediapipe==0.10.9
mistune==3.0.2
more-itertools==8.10.0
mpmath==1.3.0
mypy==1.0.1
mypy-extensions==0.4.3
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
netifaces==0.11.0
networkx==3.2.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.24.2
oauthlib==3.2.2
olefile==0.46
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
overrides==7.4.0
packaging==23.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pgzero==1.2
picamera2==0.3.14
pidng==4.0.9
piexif==1.1.3
pigpio==1.78
Pillow==9.4.0
platformdirs==2.6.0
prometheus-client==0.18.0
prompt-toolkit==3.0.41
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
PyAudio==0.2.14
pybind11==2.11.1
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydub==0.25.1
pygame==2.1.2
Pygments==2.14.0
PyGObject==3.42.2
pyinotify==0.9.6
PyJWT==2.6.0
pylint==2.16.2
PyOpenGL==3.1.6
pyOpenSSL==23.0.0
pyparsing==3.1.1
PyQt5==5.15.9
PyQt5-sip==12.11.1
pyserial==3.5
pysmbc==1.0.23
python-apt==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-engineio==4.8.0
python-json-logger==2.0.7
python-prctl==1.8.1
python-socketio==5.11.0
pyttsx3==2.7
pytz==2022.7.1
pyudev==0.24.0
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.31.0
reportlab==3.6.12
requests==2.31.0
requests-oauthlib==1.3.0
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roman==3.3
rpds-py==0.12.0
RPi.GPIO==0.7.1a4
RTIMULib==7.2.1
Send2Trash==1.8.2
sense-hat==2.4.0
simple-websocket==1.0.0
simplejpeg==1.6.6
simplejson==3.18.3
six==1.16.0
smbus2==0.4.2
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.3.2
spidev==3.5
ssh-import-id==5.10
stack-data==0.6.3
sympy==1.12
terminado==0.18.0
tflite-runtime==2.14.0
tflite-support==0.1.0a1
thonny==4.1.3
tinycss2==1.2.1
toml==0.10.2
tomlkit==0.11.7
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.3.3
traitlets==5.13.0
twython==3.8.2
types-aiofiles==22.1
types-annoy==1.17
types-appdirs==1.4
types-aws-xray-sdk==2.10
types-babel==2.11
types-backports.ssl-match-hostname==3.7
types-beautifulsoup4==4.11
types-bleach==5.0
types-boto==2.49
types-braintree==4.17
types-cachetools==5.2
types-caldav==0.10
types-certifi==2021.10.8
types-cffi==1.15
types-chardet==5.0
types-chevron==0.14
types-click-spinner==0.1
types-colorama==0.4
types-commonmark==0.9
types-console-menu==0.7
types-contextvars==2.4
types-croniter==1.3
types-cryptography==3.3
types-D3DShot==0.1
types-dateparser==1.1
types-DateTimeRange==1.2
types-decorator==5.1
types-Deprecated==1.2
types-dj-database-url==1.0
types-docopt==0.6
types-docutils==0.19
types-editdistance==0.6
types-emoji==2.1
types-entrypoints==0.4
types-first==2.0
types-flake8-2020==1.7
types-flake8-bugbear==22.10.27
types-flake8-builtins==2.0
types-flake8-docstrings==1.6
types-flake8-plugin-utils==1.3
types-flake8-rst-docstrings==0.2
types-flake8-simplify==0.19
types-flake8-typing-imports==1.14
types-Flask-Cors==3.0
types-Flask-SQLAlchemy==2.5
types-fpdf2==2.5
types-gdb==12.1
types-google-cloud-ndb==1.11
types-hdbcli==2.14
types-html5lib==1.1
types-httplib2==0.21
types-humanfriendly==10.0
types-invoke==1.7
types-JACK-Client==0.5
types-jmespath==1.0
types-jsonschema==4.17
types-keyboard==0.13
types-ldap3==2.9
types-Markdown==3.4
types-mock==4.0
types-mypy-extensions==0.4
types-mysqlclient==2.1
types-oauthlib==3.2
types-openpyxl==3.0
types-opentracing==2.4
types-paho-mqtt==1.6
types-paramiko==2.11
types-parsimonious==0.10
types-passlib==1.7
types-passpy==1.0
types-peewee==3.15
types-pep8-naming==0.13
types-Pillow==9.3
types-playsound==1.3
types-polib==1.1
types-prettytable==3.4
types-protobuf==3.20
types-psutil==5.9
types-psycopg2==2.9
types-pyaudio==0.2
types-PyAutoGUI==0.9
types-pycurl==7.45
types-pyfarmhash==0.3
types-pyflakes==2.5
types-Pygments==2.13
types-pyinstaller==5.6
types-PyMySQL==1.0
types-pynput==1.7
types-pyOpenSSL==22.1
types-pyRFC3339==1.1
types-PyScreeze==0.1
types-pysftp==0.2
types-pytest-lazy-fixture==0.6
types-python-crontab==2.6
types-python-dateutil==2.8.19.14
types-python-gflags==3.1
types-python-jose==3.3
types-python-nmap==0.7
types-python-slugify==6.1
types-pytz==2022.6
types-pyvmomi==7.0
types-pywin32==304
types-PyYAML==6.0
types-redis==4.3
types-regex==2022.10.31
types-requests==2.28
types-retry==0.9
types-Send2Trash==1.8
types-setuptools==65.5
types-simplejson==3.17
types-singledispatch==3.7
types-six==1.16
types-slumber==0.7
types-SQLAlchemy==1.4.43
types-stdlib-list==0.8
types-stripe==3.5
types-tabulate==0.9
types-termcolor==1.1
types-toml==0.10
types-toposort==1.7
types-tqdm==4.64
types-tree-sitter==0.20
types-tree-sitter-languages==1.5
types-ttkthemes==3.2
types-typed-ast==1.5
types-tzlocal==4.2
types-ujson==5.5
types-urllib3==1.26
types-vobject==0.9
types-waitress==2.1
types-whatthepatch==1.0
types-xmltodict==0.13
types-xxhash==3.0
types-zxcvbn==4.4
typing_extensions==4.4.0
uri-template==1.3.0
urllib3==1.26.12
v4l2-python3==0.3.2
wcwidth==0.2.10
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==2.2.2
widgetsnbextension==4.0.9
wrapt==1.14.1
wsproto==1.2.0
zipp==1.0.0
